Move PaymentTheme to same level as PaymentUI
