+++ example-basic/src/main/java/net/optile/example/basic/BasicActivity.java
@@ -23,7 +23,7 @@ import net.optile.payment.core.PaymentError;
 import net.optile.payment.model.Interaction;
 import net.optile.payment.ui.PaymentResult;
 import net.optile.payment.ui.PaymentUI;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 
 /**
  * This is the main Activity of this basic example app
