+++ payment/src/main/java/net/optile/payment/ui/list/AccountCardViewHolder.java
@@ -17,7 +17,7 @@ import net.optile.payment.R;
 import net.optile.payment.model.AccountMask;
 import net.optile.payment.ui.model.AccountCard;
 import net.optile.payment.ui.model.PaymentCard;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 import net.optile.payment.util.PaymentUtils;
 
 /**
