+++ payment/src/main/java/net/optile/payment/ui/list/ListAdapter.java
@@ -22,7 +22,7 @@ import net.optile.payment.ui.model.AccountCard;
 import net.optile.payment.ui.model.NetworkCard;
 import net.optile.payment.ui.model.PaymentCard;
 import net.optile.payment.ui.model.PresetCard;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 import net.optile.payment.validation.ValidationResult;
 import net.optile.payment.validation.Validator;
