+++ payment/src/main/java/net/optile/payment/ui/page/BasePaymentActivity.java
@@ -25,7 +25,7 @@ import net.optile.payment.ui.dialog.DialogHelper;
 import net.optile.payment.ui.dialog.MessageDialogFragment;
 import net.optile.payment.ui.dialog.ThemedDialogFragment;
 import net.optile.payment.ui.dialog.ThemedDialogFragment.ThemedDialogListener;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 
 /**
  * The base activity for payment activities.
