+++ payment/src/main/java/net/optile/payment/ui/widget/CheckBoxWidget.java
@@ -15,7 +15,7 @@ import android.widget.TextView;
 import net.optile.payment.R;
 import net.optile.payment.core.PaymentException;
 import net.optile.payment.form.Operation;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 
 /**
  * Widget for showing the CheckBox input element
