+++ payment/src/main/java/net/optile/payment/ui/widget/LabelWidget.java
@@ -11,7 +11,7 @@ package net.optile.payment.ui.widget;
 import android.view.View;
 import android.widget.TextView;
 import net.optile.payment.R;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 
 /**
  * Widget for showing an info label in the PaymentCard
