+++ payment/src/main/java/net/optile/payment/ui/widget/WidgetInflater.java
@@ -15,7 +15,7 @@ import android.view.ViewGroup;
 import net.optile.payment.R;
 import net.optile.payment.model.InputElement;
 import net.optile.payment.model.InputElementType;
-import net.optile.payment.ui.theme.PaymentTheme;
+import net.optile.payment.ui.PaymentTheme;
 
 /**
  * Class with helper methods to inflate widgets with proper theming
